requirements.txt
asgiref==3.10.0
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf_yasg==1.21.11
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
python-decouple==3.8
requests==2.32.5
stripe==14.0.1
Pillow==12.0.0